Animal: eat()

+++ src/buildings/Barn.java
@@ -8,8 +8,8 @@ public class Barn extends Building{
         super("Stodoła", "Rośliny", 20000, 500.00, 400.00);
     }
 
-    public Integer kgOfPlantsStoredInBarn(Farm farm){
-        Integer allPlantsInKg = farm.getCorn().getStoredInKg() + farm.getLatePotato().getStoredInKg() + farm.getLateRape().getStoredInKg() + farm.getLateRye().getStoredInKg() + farm.getSpringBarley().getStoredInKg() + farm.getWhiteBeet().getStoredInKg() + farm.getWinterWheat().getStoredInKg();
+    public Double kgOfPlantsStoredInBarn(Farm farm){
+        Double allPlantsInKg = farm.getCorn().getStoredInKg() + farm.getLatePotato().getStoredInKg() + farm.getLateRape().getStoredInKg() + farm.getLateRye().getStoredInKg() + farm.getSpringBarley().getStoredInKg() + farm.getWhiteBeet().getStoredInKg() + farm.getWinterWheat().getStoredInKg();
         return allPlantsInKg;
     }
